squash old migrations 201806
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/342061
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Migration-Review: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>

+++ lib/tasks/db.rake
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 Rake::Task["db:drop"].enhance do
+  next unless Canvas.redis_enabled?
+
   GuardRail.activate(:deploy) { Canvas.redis.flushdb }
 end
